+++ esa-20240910/ChangeLog.md
@@ -1,3 +1,9 @@
+2025-01-10 Version: 2.10.2
+- Update API ListPages: add param QueryArgs.
+- Update API ListUserRatePlanInstances: update response param.
+- Update API SetCertificate: update response param.
+
+
 2025-01-09 Version: 2.10.1
 - Generated python 2024-09-10 for ESA.